Provide a GF to bind specials during inspection of data.
When lots of objects are used, memory consumption is reduced by
keeping a reference to some higher-level object in a special,
ie. to not duplicate it everywhere.

For example, overview objects store an array of files;
and then there are lots of records that need to reference the
overview object and the file. Only storing the (small 1-byte)
index in the file array in a record makes them smaller;
the application uses *overview* everywhere to allow derefencing.


But that means that the small indizes that swank produces on INSPECT
don't really help that much - they need to be resolved!

So my code contains forms like these:

  (defmethod swank:on-inspect-bind-to ((x overview))
    '*overview*)

  (defmethod swank:emacs-inspect :around ((r record))
    (swank/backend:label-value-line*
      ('file (aref (file-index r)
                   (files *overview*)))
      ...))